Extract HiddenFocusStartPoint as it's own file
